refactor: better organize types and configs
