+++ src/configs/indentation.ts
@@ -3,6 +3,8 @@
 export const indentation = {
   test: '  ',
   stdio: '      ',
+  describeCounter: 0,
+  testCounter: 0,
 };
 
 /* c8 ignore stop */
