+++ src/index.ts
@@ -2,7 +2,8 @@
 export { poku } from './modules/poku.js';
 export { exit } from './modules/exit.js';
 export { assert } from './modules/assert.js';
-export { describe, log } from './modules/describe.js';
+export { describe } from './modules/describe.js';
+export { log } from './modules/log.js';
 export { assertPromise } from './modules/assert-promise.js';
 export { beforeEach, afterEach } from './modules/each.js';
 export { publicListFiles as listFiles } from './modules/list-files-sync.js';
